refactored network settings
