+++ core/conf/conf_network.h
@@ -1,0 +1,9 @@
+#ifndef CONF_NETWORK_H
+#define CONF_NETWORK_H
+
+#define SSID = "your_ssid";
+#define SSID_PW = "your_password";
+#define AGENT_IP   "your_adgent_ip" // IP of the ROS2 messages recceiver
+#define AGENT_PORT 2018 //AGENT port number
+
+#endif
